+++ GoRogue.UnitTests/FOV/FOVSenseMapEquivalenceTests.cs
@@ -4,6 +4,7 @@ using GoRogue.FOV;
 using GoRogue.Random;
 using GoRogue.SenseMapping;
 using GoRogue.UnitTests.Mocks;
+using GoRogue.UnitTests.SenseMapping;
 using SadRogue.Primitives;
 using SadRogue.Primitives.GridViews;
 using ShaiRandom.Generators;
@@ -63,7 +64,7 @@ namespace GoRogue.UnitTests.FOV
             var senseMap = new SenseMap(s_resMap);
 
             // Set up sense source (using shadow-casting to match LOS)
-            var lightSource = new SenseSource(SourceType.Shadow, source, Radius, shape);
+            var lightSource = AlgorithmFactory.CreateSenseSource(SourceType.Shadow, source, Radius, shape);
             senseMap.AddSenseSource(lightSource);
 
             // Calculate LOS and sense map
@@ -72,12 +73,12 @@ namespace GoRogue.UnitTests.FOV
 
             // Verify equivalence of LOS and SenseMap
             Assert.Equal(Width, fov.DoubleResultView.Width);
-            Assert.Equal(Width, senseMap.Width);
+            Assert.Equal(Width, senseMap.ResultView.Width);
             Assert.Equal(Height, fov.DoubleResultView.Height);
-            Assert.Equal(Height, senseMap.Height);
+            Assert.Equal(Height, senseMap.ResultView.Height);
 
             foreach (var pos in fov.DoubleResultView.Positions())
-                Assert.Equal(fov.DoubleResultView[pos], senseMap[pos]);
+                Assert.Equal(fov.DoubleResultView[pos], senseMap.ResultView[pos]);
         }
 
 
@@ -89,13 +90,9 @@ namespace GoRogue.UnitTests.FOV
             var senseMap = new SenseMap(s_resMap);
 
             // Set up sense source (using shadow-casting to match LOS)
-            var lightSource =
-                new SenseSource(SourceType.Shadow, source, Radius, shape)
-                {
-                    IsAngleRestricted = true,
-                    Angle = angle,
-                    Span = span
-                };
+            var lightSource = AlgorithmFactory.CreateSenseSource(SourceType.Shadow, source, Radius, shape);
+            lightSource.Angle = angle;
+            lightSource.Span = span;
             senseMap.AddSenseSource(lightSource);
 
             // Calculate LOS and sense map
@@ -104,12 +101,12 @@ namespace GoRogue.UnitTests.FOV
 
             // Verify equivalence of LOS and SenseMap
             Assert.Equal(Width, fov.DoubleResultView.Width);
-            Assert.Equal(Width, senseMap.Width);
+            Assert.Equal(Width, senseMap.ResultView.Width);
             Assert.Equal(Height, fov.DoubleResultView.Height);
-            Assert.Equal(Height, senseMap.Height);
+            Assert.Equal(Height, senseMap.ResultView.Height);
 
             foreach (var pos in fov.DoubleResultView.Positions())
-                Assert.Equal(fov.DoubleResultView[pos], senseMap[pos]);
+                Assert.Equal(fov.DoubleResultView[pos], senseMap.ResultView[pos]);
         }
     }
 }